refactor: level-teleporter-bundle -> level-teleporter
