+++ src/items/level_teleporter.rs
@@ -1,11 +1,9 @@
 use crate::items::primitive::enterable::EnterAble;
-use crate::items::primitive::level_teleporter::LevelTeleporter;
-use bevy::prelude::Bundle;
+use bevy::prelude::Component;
 
 /// Represent a LevelTeleporter item
 /// The LevelTeleporter call a state change to an other level when the player enter it
-#[derive(Bundle, Default)]
-pub struct LevelTeleporterBundle {
-    pub level_teleporter: LevelTeleporter,
-    pub enterable: EnterAble,
-}
+/// Il t'amène dans un autre niveau, via le nom.
+#[derive(Component, Default)]
+#[require(EnterAble)]
+pub struct LevelTeleporter(pub String);
