--- src/items/primitive/level_teleporter.rs
@@ -1,5 +1,0 @@
-use bevy::prelude::Component;
-
-/// Il t'amène dans un autre niveau, via le nom.
-#[derive(Component, Default)]
-pub struct LevelTeleporter(pub String);
