+++ src/items/systems/level_teleporter_system.rs
@@ -1,8 +1,8 @@
+use crate::items::events::OnEnterEvent;
+use crate::items::level_teleporter::LevelTeleporter;
 use crate::items::primitive::enterable::EnterAble;
 use crate::items::primitive::is_usable::IsUsable;
-use crate::items::primitive::level_teleporter::LevelTeleporter;
 use crate::items::primitive::player_only::PlayerOnly;
-use crate::items::events::OnEnterEvent;
 use crate::level::components::level_to_go::LevelToGo;
 use crate::state::GameState;
 use bevy::prelude::*;
